Clean-up tests
* Remove init files under test directories
* Move test helpers from utils files into bokeh.util.testing.py
* Fix up imports - no relative imports in tests
* Rename test files to be test_ (ensures that running bokeh.test() runs all tests)
