+++ bokeh/charts/builder/tests/test_horizon_builder.py
@@ -22,8 +22,8 @@ from numpy.testing import assert_array_equal
 import pandas as pd
 
 from bokeh.charts import Horizon
+from bokeh.util.testing import create_chart
 
-from bokeh.charts.builder.tests._utils import create_chart
 
 #-----------------------------------------------------------------------------
 # Classes and functions
