+++ bokeh/charts/tests/test_chart_utils.py
