+++ bokeh/server/tests/test_cow.py
@@ -9,18 +9,19 @@ from __future__ import absolute_import
 
 from bokeh.exceptions import AuthenticationException
 from bokeh.models.plots import Plot
-from bokeh.session import TestSession
+from bokeh.session import TestSession as Session
 from werkzeug.exceptions import Unauthorized
 
-from . import test_utils
-from ..app import app, bokeh_app
-from ..models.user import User, new_user
-from ..models.docs import Doc
-from ..serverbb import BokehServerTransaction
-from ..views.bbauth import handle_auth_error
-from ..views.main import _makedoc
+from bokeh.server.app import app, bokeh_app
+from bokeh.server.models.user import User, new_user
+from bokeh.server.models.docs import Doc
+from bokeh.server.serverbb import BokehServerTransaction
+from bokeh.server.views.bbauth import handle_auth_error
+from bokeh.server.views.main import _makedoc
+from bokeh.util.testing import FlaskClientTestCase
 
-class AuthTestCase(test_utils.FlaskClientTestCase):
+
+class AuthTestCase(FlaskClientTestCase):
     options = {'multi_user' : True}
     def test_handle_auth_error_decorator(self):
         @handle_auth_error
@@ -71,7 +72,7 @@ class AuthTestCase(test_utils.FlaskClientTestCase):
             self.assertRaises(AuthenticationException, BokehServerTransaction,
                               user, doc, 'rw', temporary_docid="foobar")
 
-class TransactionManagerTestCase(test_utils.FlaskClientTestCase):
+class TransactionManagerTestCase(FlaskClientTestCase):
     options = {'multi_user' : True}
     def setUp(self):
         super(TransactionManagerTestCase, self).setUp()
@@ -152,10 +153,10 @@ class TransactionManagerTestCase(test_utils.FlaskClientTestCase):
             t = self.transaction(None, mode='r')
             self.assertRaises(AuthenticationException, t.load, gc=True)
 
-class PublishTestCase(test_utils.FlaskClientTestCase):
+class PublishTestCase(FlaskClientTestCase):
     options = {'multi_user' : True}
     def test_publish(self):
-        sess = TestSession(client=app.test_client())
+        sess = Session(client=app.test_client())
         sess.register('testuser', 'testpassword')
         sess.use_doc('test_cow')
         sess.publish()
@@ -164,10 +165,10 @@ class PublishTestCase(test_utils.FlaskClientTestCase):
         assert doc.published == True
 
     def test_publish_fails_for_invalid_auth(self):
-        sess = TestSession(client=app.test_client())
+        sess = Session(client=app.test_client())
         sess.register('testuser', 'testpassword')
         sess.use_doc('test_cow')
-        sess2 = TestSession(client=app.test_client())
+        sess2 = Session(client=app.test_client())
         sess2.register('testuser2', 'testpassword')
         sess2.docid = sess.docid
         self.assertRaises(Exception, sess2.publish)
