+++ bokeh/server/tests/test_multiuser_auth.py
@@ -9,21 +9,22 @@ from __future__ import absolute_import
 
 import tempfile
 
-from bokeh.session import TestSession
+from bokeh.session import TestSession as Session
 
-from . import test_utils
-from ..app import bokeh_app, app
-from ..models import user
+from bokeh.server.app import bokeh_app, app
+from bokeh.server.models import user
+from bokeh.util.testing import FlaskClientTestCase
 
-class TestMultiUserAuth(test_utils.FlaskClientTestCase):
+
+class TestMultiUserAuth(FlaskClientTestCase):
     options = {'multi_user' : True}
     def test_register(self):
-        sess = TestSession(client=app.test_client())
+        sess = Session(client=app.test_client())
         sess.register('testuser1', 'testpassword1')
         sess.userinfo['username'] == 'testuser1'
 
     def test_register_twice_should_fail(self):
-        sess = TestSession(client=app.test_client())
+        sess = Session(client=app.test_client())
         sess.register('testuser1', 'testpassword1')
         self.assertRaises(RuntimeError, sess.register, 'testuser1', 'testpassword1')
 
@@ -31,7 +32,7 @@ class TestMultiUserAuth(test_utils.FlaskClientTestCase):
         username = "testuser2"
         password = "fluffy"
         user.new_user(bokeh_app.servermodel_storage, username, password)
-        sess = TestSession(client=app.test_client())
+        sess = Session(client=app.test_client())
         #have not logged in yet, so userinfo should return http error
         self.assertRaises(Exception, lambda : sess.userinfo)
 
@@ -46,20 +47,18 @@ class TestMultiUserAuth(test_utils.FlaskClientTestCase):
         #create a dummy config file
         config = tempfile.mkdtemp()
         #create server
-        server = TestSession(name="foo",
-                             client=app.test_client(),
-                             configdir=config,
-                             load_from_config=True
-        )
+        server = Session(name="foo",
+                         client=app.test_client(),
+                         configdir=config,
+                         load_from_config=True)
         assert server.userapikey == "nokey"
         #register a user
         server.register("testuser", "fluffy")
         assert server.userapikey and server.userapikey != "nokey"
-        server2 = TestSession(name="foo",
-                              client=app.test_client(),
-                              configdir=config,
-                              load_from_config=True
-        )
+        server2 = Session(name="foo",
+                          client=app.test_client(),
+                          configdir=config,
+                          load_from_config=True)
         assert server2.userapikey == server.userapikey
         assert server2.root_url == server.root_url
 
@@ -69,21 +68,19 @@ class TestMultiUserAuth(test_utils.FlaskClientTestCase):
     def test_login2(self):
         #create a server config, register a user
         config1 = tempfile.mkdtemp()
-        server = TestSession(name="foo",
-                             client=app.test_client(),
-                             configdir=config1,
-                             load_from_config=True
-        )
+        server = Session(name="foo",
+                         client=app.test_client(),
+                         configdir=config1,
+                         load_from_config=True)
         assert server.userapikey == "nokey"
         server.register("testuser", "fluffy")
         assert server.userapikey and server.userapikey != "nokey"
 
         config2 = tempfile.mkdtemp()
-        server2 = TestSession(name="foo",
-                             client=app.test_client(),
-                             configdir=config2,
-                             load_from_config=True
-        )
+        server2 = Session(name="foo",
+                          client=app.test_client(),
+                          configdir=config2,
+                          load_from_config=True)
         assert server2.userapikey == "nokey"
         server2.login("testuser", "fluffy")
         #make sure the userapikeys match
