+++ bokeh/util/tests/test_serialization.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 import unittest
 
-from bokeh.util.serialization import get_json, json_apply, make_id, urljoin
+from bokeh.util.serialization import get_json, json_apply, make_id, urljoin, traverse_data
 
 class DummyRequestCallable():
     def json(self):
@@ -58,6 +58,20 @@ class TestJsonapply(unittest.TestCase):
         self.assertEqual(result, [[['goaled', 'junk'], 'junk', 'junk']])
         result = json_apply({'1': 'goal', 1.5: {'2': 'goal', '3': 'junk'}}, check_func, func)
         self.assertEqual(result, {'1': 'goaled', 1.5: {'2': 'goaled', '3': 'junk'}})
+
+
+class TestTraverseData(unittest.TestCase):
+    testing = [[float('nan'), 3], [float('-inf'), [float('inf')]]]
+    expected = [['NaN', 3.0], ['-Infinity', ['Infinity']]]
+
+    def test_return_valid_json(self):
+        self.assertTrue(traverse_data(self.testing) == self.expected)
+
+    def test_with_numpy(self):
+        self.assertTrue(traverse_data(self.testing, True) == self.expected)
+
+    def test_without_numpy(self):
+        self.assertTrue(traverse_data(self.testing, False) == self.expected)
 
 if __name__ == "__main__":
     unittest.main()
